reSWIGged
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_6_BRANCH@39910 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ wxPython/src/msw/_gdi_wrap.cpp
@@ -1977,8 +1977,8 @@ static wxBitmap *new_wxBitmap(PyObject *listOfStrings){
             return bmp;
         }
 static wxBitmap *new_wxBitmap(PyObject *bits,int width,int height,int depth=1){
-            char* buf;
-            int   length;
+            char*       buf;
+            Py_ssize_t  length;
             PyString_AsStringAndSize(bits, &buf, &length);
             return new wxBitmap(buf, width, height, depth);
         }
